rocky/requirements.txt
ares==0.7.1
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.12.7
chardet==5.1.0
cwe==1.6
Django==3.2.18
django-formtools==2.4
django-otp==1.1.3
django-phonenumber-field==6.1.0
django-two-factor-auth==1.14.0
idna==3.4
lxml==4.9.2
phonenumbers==8.12.57
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
pytz==2022.6
qrcode==7.3.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
toolz==0.12.0
typing-extensions==4.4.0
urllib3==1.26.14
django-markdownify==0.9.2
xhtml2pdf==0.2.8
dnspython~=2.1.0
django-password-validators==1.7.0
colorama==0.4.6
django-csp==3.7
uwsgi==2.0.21
djangorestframework==3.14.0
django-tagulous==1.3.3
drf-standardized-errors==0.12.5

# temp fix to pass build, remove later when https://github.com/xhtml2pdf/xhtml2pdf/issues/589 is solved
reportlab==3.6.12
